requirements.txt
numpy>=1.19.2
pandas>=1.2.0
tensorflow>=2.4.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
openai>=1.3.0
ollama>=0.1.4
python-dotenv>=1.0.0
shap>=0.41.0
gradio>=4.0.0
plotly>=5.18.0
